QB30-021 Discriminants and components in protected functions
In Get_Variables we used to return as variables all the discriminants,
components and Part_Of of a protected type. We now don't do that if the
scope in which they are used is a protected function as the current
instance of the protected type acts as an IN parameter.

* gnat2why/flow/flow_utility.adb
(Do_Entity): do not consider as variable a discriminant, a component or a
Part_Of of a protected type used within a protected function.

* testsuite/gnatprove/tests/QB30-021__flow_prot_fun
Extend test to consider expressions initialized through a Part_Of the
protected object. Update test output. Spurious errors are now gone.

Change-Id: I1d14618c4852377ebc38bbe5d6f5c986a2f2975b