P517-019 Enable test that was XFAIL
Test Q925-025 can now be enabled as we are correctly detecting expression
with variable input in a subtype constraint.
Most likely this has been fixed in a recent commit for P517-019 but we
couldn't notice as the test was XFAIL.

* testsuite/gnatprove/tests/Q925-025__flow_type_bound
Enable test and add test output. We now detect expression with variable
inputs.

Change-Id: I4ee3d93e0a257299675554cd04d19e37d411e71f