QC12-068 Hidden states in private children
The traversal to find unexposed hidden state was not diving into the hidden
part of a private package (private and body part). We are now traversing
those parts to find any hidden unexposed state. We also failed to detect
constants with variable inputs declared in a package body. We now do
correctly detect them.

* gnat2why/flow/flow-analysis.adb
(Find_Hidden_Unexposed_State): we now traverse into the private and the
body part of a private child in case it doesn't have an abstract state but
its parent package does. This is to find any unexposed hidden states.
(Check_Hidden_State_Variables_And_Missing_Part_Of): we now emit a check
for hidden state that appears in the private or body part of a private
child. We also correctly detect constants with variable inputs declared
in package bodies.

* testsuite/gnatprove/QC12-068__flow_priv_child
Expand test and update output with checks that were previously missing.

* testsuite/gnatprove/QC14-061__flow_hidden_state
Test output updated. We now detect a constant with variable inputs in the
package body.

Change-Id: Id9e3606dfd727f96c170f3f254ac08084a37ba5d