Q908-006 Read of uninitialized variable from interrupt handler
If we read an uninitialized variable from an interrupt handler we now emit
an error as we do for task types and main programs.

* gnat2why/flow/flow.ads
(Flow_Analyse_Entity): we now set FA.Main to true if the subprogram is an
interrupt handler.

* testsuite/gnatprove/tests/Q908-006__flow_global_handler
Test output updated. We now emit an error because the interrupt handler
reads an uninitialized variable.

Change-Id: I446b1f1c40128ea6ed1b486408e56e5d647edff9