don't check for missing stubs in the backend
The (now deconstructed) traversal for frontend cross-references was called
very early in the frontned and had to deal with missing stubs. Its code was
copy-pasted and adapted as a traversal for generated globals, which is only
called in the backend with all stubs filled-in by their source files (i.e.
if stubs are missing then the backend is not called at all).

This patch deconstructs detection of missing stubs.

(no-tn-check)

* gnat2why/flow/flow_generated_globals-traversal.adb
(Traverse_Stub): remove, as it is not needed in the backend.

Change-Id: I71756193b0bf324303006553c8c87c05399a4a91