Q515-016 Add ghost status to the ALI file
We are now recording in the ALI file the ghost status of objects. This is
in turn very useful when checking SPARK RM 6.9(20) when we want to detect
if a ghost procedure has any non-ghost outputs. Such non-ghost global
outputs can sometimes only be known by Entity_Name (as in the test for
Q918-039) and so far we didn't have a way to retrieve their ghost status.

* gnat2why/flow/flow_generated_globals-ali_serialization.ads
* gnat2why/flow/flow_generated_globals-ali_serialization.adb
* gnat2why/flow/flow_generated_globals-phase_1.adb
* gnat2why/flow/flow_generated_globals-phase_2.ads
* gnat2why/flow/flow_generated_globals-phase_2.adb
Write ghost status for objects into the ALI file

* gnat2why/flow/flow_utility.adb
(Is_Ghost_Object): when asking for the ghost status of an object known by
Entity_Name we now check what is stored in the ALI file.

* testsuite/gnatprove/tests/Q918-039__flow_ghost_entity_names
Test output updated. We now don't emit a spurious check when a ghost
procedures has a ghost global output only known by Entity_Name. Before we
were always returning a ghost status of FALSE for every Entity_Name object.

Change-Id: I03a5613fe8e72c4a6cb2b532efb0b22f825d9710