minor rename Is_Ghost_Object to Is_Ghost_Entity
This renaming is motivated by a recent commit under Q918-039 for which we
are now detecting the ghost status for all entities (including abstract
states). Before Is_Ghost_Object was only caring about the ghost status of
objects, but abstract states are not objects so we needed to extend the
function to be Is_Ghost_Entity to deal with them.

(no-tn-check)

Change-Id: I6f1ba61e7dfe21557d7cd91a727bbd9eed3503ac