QA13-044 Detect discriminant constraint in Get_Variables
We failed to detect discriminant constraints used in the declaration of a
variable. We now detect them and return them when calling Get_Variables.

* gnat2why/flow/flow_utility.adb
(Do_Entity): if the variable is of a record type, we check if it has
discriminant constraints and if so we return them as variables used.

* testsuite/gnatprove/tests/PB02-004__inherited_sub
Test output updated. We now no longer issue spurious warnings.

* testsuite/gnatprove/tests/QA13-044__flow_discirminant
Test output updated. We now no longer issue spurious warnings and checks.

Change-Id: Ic9116093668221e5d4db11a3116e5119fd9c7fc6