Q912-041 ignore calls when explicit Global is given
For subprograms with explicit Global, SPARK_Mode => On for spec and
SPARK_Mode => Off for body, we shall ignore proof/conditional/definite calls
and simply trust the Global contract. This restores the behaviour of the old
GG; it was changed by a mistake in the new GG (as the only tests where it
matters are enough complicated to make a mistake in manual diagnosis).

For simplicity we do the same were both spec and body has SPARK_Mode => Off.

Also, remove assertion relating calls in the Pre/Post contracts discovered
by the flow analysis and those discovered by the frontend cross-references,
because the latter are well-known to be unreliable (and nearly impossible to
fix).

* gnat2why/flow/flow_generated_globals-partial.adb
(Preanalyze_Spec): fix behaviour for entities with mixed SPARK_Mode; remove
 unreliable assertion.

* testsuite/gnatprove/tests/P718-001__ubstr_bad/test.out
* testsuite/gnatprove/tests/P718-001__ubstr_good/test.out
Update outputs (reverts changes made by the new GG).

* testsuite/gnatprove/tests/N709-001__contracts/test.opt
* testsuite/gnatprove/tests/O916-043__crazyflie/test.opt
* testsuite/gnatprove/tests/O916-052__crazyflie/test.opt
* testsuite/gnatprove/tests/O917-119__crazyflie/test.opt
* testsuite/gnatprove/tests/P725-016__recursive_spec/test.opt
Reenable; they no longer crash on assertion related to the frontend
cross-references.

Change-Id: I966d59aeb4eb2c914a38f2ab88420023a6ef9256