QC19-031 Restore missing checks on inlined subprograms
The inlining done in GNATprove on local subprograms without contracts
could cause a check to be skipped on the returned value. The fix here
completes the fix in GNAT frontend inlining mechanism.

* gnat2why/why/gnat2why-expr.adb
(Transform_Attr): Set the target type to the base type of the prefix of
 Mod attribute.

* testsuite/gnatprove/tests/K303-015__default/test.out
* testsuite/gnatprove/tests/M924-029__discr_dyn/test.out
* testsuite/gnatprove/tests/N317-015__inlining/test.out
* testsuite/gnatprove/tests/P825-006__inlined_range/test.out
* testsuite/gnatprove/tests/autopilot/test.out
Update outputs.

* testsuite/gnatprove/tests/QC19-031__mod_attribute/test.opt
* testsuite/gnatprove/tests/QC19-031__mod_attribute/test.out
Test passes now, by issuing an error.

Change-Id: Icd38086fabd3714e609766915692d4192c7dda9f