R111-038 Enforce SPARK RM 6.1.4(20)
GNATprove was not enforcing the second part of SRM 6.1.4(20) for which a
subprogram or entry having such input or proof_in global items that are
marked as costants after elaboration shall not be called during library
unit elaboration. Also, the improved check now works also for variables
that are only known by entity names.

* gnat2why/flow/flow-analysis.ads
* gnat2why/flow/flow-analysis.adb
(Check_Constant_After_Elaboration): check package elaborations

* gnat2why/flow/flow.adb
(Flow_Analyse_CUnit): call Check_Constant_After_Elaboration for packages

* gnat2why/flow/flow_generated_globals-ali_serialization.ads
* gnat2why/flow/flow_generated_globals-ali_serialization.adb
* gnat2why/flow/flow_generated_globals-phase_1.adb
* gnat2why/flow/flow_generated_globals-phase_2.ads
* gnat2why/flow/flow_generated_globals-phase_2.adb
Code to add the CAE status of a variable into the ALI file and then read it
in phase 2.

* gnat2why/flow/flow_utility.ads
* gnat2why/flow/flow_utility.adb
(Is_Constant_After_Elaboration): routine that works for Flow_Ids

* testsuite/gnatprove/tests/O429-038__flow_check_cae
Test output updated. We now reference the SRM and we emit a new check that
we missed before.

* testsuite/gnatprove/tests/OA23-016__flow_local_var_at_lhs_of_gen_init
Test output updated. We now emit two checks that we missed before.

* testsuite/gnatprove/tests/R111-038__flow_cae_elab
Test output updated. We now emit a check for rule 6.1.4(20)

* testsuite/gnatprove/tests/R111-038__flow_cae_entity_names
Add test to show the new code works for entity names. We correctly detect
that Q has as a global output a constant after elaboration.

Change-Id: I79ea1994aad6f4fc0cbbd0472cd112c8e3d36e71
Signed-off-by: Elisa Barboni <elisa.barboni@altran.com>