R122-003 better error locations for crashes in SPARK_Register
The code for registering mapping between Entity_Id and Entity_Name has a bad
history of crashes on atypical source code. Set the error location on entry/
exit, so that when it crashes, the bug box will point to the offending code.

* gnat2why/spark/spark_register.adb
(Register_Compilation_Unit): set/restore Current_Error_Node when traversing
 the compilation unit.

Change-Id: Id45872a34654cd5fee89c9add553c8fb8c2de4da