Q317-010 Respect SPARK_Mode when flattening a variable
When flattening variables, flow analysis was diving into parts hidden
behind a SPARK_Mode => Off. We now respect the SPARK boundary.

* gnat2why/flow/flow_utility.adb
(Components): only return discriminants and components that are in SPARK
(Flatten_Variable): we now do not flatten the variable if it is not in
SPARK.

* gnat2why/spark/spark_definition.ads
* gnat2why/spark/spark_definition.adb
Renamed the iterator Marked_Entities to SPARK_Entities as it now iterates
over Entities_In_SPARK.

* testsuite/gnatprove/tests/N609-025__private_type_discr
Test output updated. There is a spurious warning for which I will open a
new ticket.

* testsuite/gnatprove/tests/N624-030__funcs_on_containers
* testsuite/gnatprove/tests/NB14-023__DIC_on_formal_containers
* testsuite/gnatprove/tests/P715-019__flow_generic_termination
Test output updated. We don't emit anymore info messages on non visible
components of an object.

* testsuite/gnatprove/tests/M607-025__rev_stack
* testsuite/gnatprove/tests/M611-052__discr
* testsuite/gnatprove/tests/MA07-042__flow_formal_container_hack
* testsuite/gnatprove/tests/N110-032__list_conversion
Test output updated. The current output is a consequence of a bug in
marking (R110-037). The reason why we now have the message on the object
is because in the flatting of the variable we wrongly return False when
we call Entity_In_SPARK and this will stop the flattening and just return
the object. Before we didn't respect the SPARK_Mode boundary and we would
flatten the variable and represent its private part through "private part".

* testsuite/gnatprove/tests/O403-005__private_attributes
* testsuite/gnatprove/tests/P718-001__ubstr_good
Test output updated. We have an extra info message coming from the fact
that the variables are of a private record type with private part hidden
behind a SPARK_Mode => Off. For this reason we cannot see its components
and refer to its "private part".

Change-Id: Ic5d2896de5117ac18988bab338e0837a00448b08