Q317-010 unify processing of external axiom types in flow and proof
The SPARK_Definition.SPARK_Entities iterator was only exposed for flow to
get both "entities to translate" and those in external axioms. It turns out
that in proof we already had a way to iterate on both of these.

This patch merely reuses the proof iteration, thus eliminating the need for
the SPARK_Entities iterator. The code should become marginally faster too,
because we avoid branching at each call to the iteration routine (to check
whether we iterate over Entities_To_Translate or SPARK_Entities).

Semantics unchanged, so no tests affected.

* gnat2why/flow/flow_utility.adb
(Initialize): rewrite using Entities_To_Translate (for types in SPARK) and
 Process_External_Entities (for types with external axioms).

(Process): refactored into a dedicated procedure; needed to call the
 Process_External_Entities routine.

* gnat2why/spark/spark_definition.ads
* gnat2why/spark/spark_definition.adb
(SPARK_Entities): remove enum and simplify the related iterators.

Change-Id: I9b78a8dcb91621f0049ee3b70a12344f778df64a