R205-008 fix slocs in messages about wrong flow dependencies
This patch splits the Search_Contract routine into Search_Depends_Contract
and Search_Initializes_Contract, which now traverse the contract explicitly
and not with the generic Traverse_Func routine. This makes them easier to
understand and diagnose; in particular, the new code does not need to check
the node kind of the parent node to know what to look for (which was broken
in the old code).

Also, for qualified names like "X.Y.Z" the new code returns the node of "X",
while the old code was returning the node of the dot between "Y" and "Z".

Finally, it deals with the current instances of single concurrent units,
which in flow are represented by a type while in contract they are
represented by an object. (This effectively reverses the transformations
done in Parse_Raw_Dependency_Map).

* gnat2why/flow/flow_utility.ads
* gnat2why/flow/flow_utility.adb
(Search_Contract): rewrite and split into dedicated routines for
 [Refined_]Depends and Initializes.

(Search_Depends_Contract): rewritten from Search_Contract.
(Search_Initializes_Contract): rewritten from Search_Contract.

* gnat2why/flow/flow_classwide.adb
* gnat2why/flow/flow-analysis.adb
Use dedicated routines for Depends/Initializes contracts. Also, when
flagging constants without variable input from Check_Depends_Contract put
the message precisely on the constant and not on the contract.

* testsuite/gnatprove/tests/M314-014__flow_packages/test.out
* testsuite/gnatprove/tests/MA11-028__flow_compute_refined_global_depends/
* testsuite/gnatprove/tests/MA15-019__flow_contract_check_initializes/...
* testsuite/gnatprove/tests/NA15-048__flow_better_error_locations/test.out
* testsuite/gnatprove/tests/O206-030__generic/test.out
* testsuite/gnatprove/tests/O324-021__trailing_R/test.out
* testsuite/gnatprove/tests/O414-015__flow_constants_in_contracts/test.out
* testsuite/gnatprove/tests/O414-015__flow_constants_in_contracts...
* testsuite/gnatprove/tests/OC08-044__refinement/test.out
* testsuite/gnatprove/tests/Q306-006__flow_no_gg/test.out
* testsuite/gnatprove/tests/Q707-013__flow_wrong_generated_initializes/...
* testsuite/gnatprove/tests/TU__depends_illegal/test.out
* testsuite/gnatprove/tests/TU__types_and_subtypes_illegal/test.out
* testsuite/gnatprove/tests/thumper/test.out
Update slocs in the error messages.

Change-Id: I5764c86314995c4f66abce7c672fe7c04c75bc04