OB17-029 Collect directly called POs from elaborations
In phase 2 we now collect POs that are directly called in elaborations of
main subprograms so that we can enforce the ceiling priority protocol check
for those.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Directly_Called_Protected_Objects): return also the POs accessed by the
withed units.

* testsuite/gnatprove/tests/OB17-029__ceiling_elaboration
* testsuite/gnatprove/tests/OB17-029__ceiling_elaboration_generics
Test output updated. We now detect the failure of the check related to the
ceiling priority protocol.

* testsuite/gnatprove/tests/OB17-029__ceiling_elaboration_task
Edit test in order not to have spurious messages on procedure P. Test
output updated. We now detect the failure of the check related to the
ceiling priority protocol.

Change-Id: Idb8e6eccaa24ab17c31d529f69062bc34c08b64a