R202-006 Stabilize gnat_server test
Removed exit request that could be triggered before all notifications are
done. Anonymise the number given to the new node. It was possible that the
request sent for executing z3 executes before the one for
introduce_premises because, in gnat_server, we now executes timeout
(provers) before idles (transformations) so the prover node is created
before the transformation node.
The node id of the node below introduce_premises being now impredictible,
we removed the get task for this node.

* testsuite/gnatprove/tests/QA11-003__itp_server_test/test.in
Removed Exit and second get_task

* testsuite/gnatprove/tests/QA11-003__itp_server_test/test.py
Exit not present anymore so we rely on timeout. children_88 and 89 created
for the children of this given node: we count them.

* testsuite/gnatprove/tests/QA11-003__itp_server_test/test.out
Update output.

Change-Id: I68acb37776ff0bde090da1cf8492385912af1142