R122-008 Remove Iterate_Generic_Parameters
With this patch we have removed the generic Iterate_Generic_Parameters as
it is not strictly necessary and the implementation had some bugs. Because
it was only called in two places we have now rewritten the code in those
routines which now looks cleaner and easier.

* gnat2why/flow/flow-analysis-sanity.adb
(Check_Actuals): rewritten

* gnat2why/spark/spark_util.ads
* gnat2why/spark/spark_util.adb
(Iterate_Generic_Parameters): deleted
(Generic_Actual_Subprograms): rewritten

* testsuite/gnatprove/tests/R122-008__flow_iterate_generic_parameters
Add test

Change-Id: I70ecda4a2d0b7ff69a01a44c471eae08580efb51