QA13-044 Detect discriminant constraint for constants
We failed to detect discriminant constraints used in the declaration of a
constant. We now detect them and return them when calling Get_Variables.

* gnat2why/flow/flow_utility.adb
(Discriminant_Constraints): new routine that returns the discriminant
constraints.
(Do_Entity): if the constant is of a record or concurrent type, we check
if it has discriminant constraints and if so we return them as variables
used.

* testsuite/gnatprove/tests/QA13-044__flow_discirminant
Test modified to include constants and selected components. Test output
updated. We now detect use of discriminants in constants but still need a
fix for selected components. This will be addressed in the next patch.

Change-Id: I7a759275a5bfb47d5340ba7ca2d014b39bbc7b10