Q929-022 Avoid generating spurious checks on type declarations
Type declarations which only rename an existing type do not introduce
constraints, and thus checking the inherited constraints for such
declarations is useless. In some cases, this leads to spurious
unproved checks, as the necessary assumptions related to the constraints
are not present in the context.

This concerns both subtype declarations:
  subtype Sub is Existing;
as well as derived type declarations:
  type Der is new Existing;

* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): Do not generate checks for constraints on
 type declarations, when these declarations only rename an existing
 type without intoducing any new constraint.

* testsuite/gnatprove/tests/Q929-022__flow_illegal_varinput/test.out
Update output, test now generates error.

* .../O414-015__flow_constants_in_contracts_review_tests/test.out
* testsuite/gnatprove/tests/Q106-015__default_init/test.out
* testsuite/gnatprove/tests/Q106-022__default_value/test.out
* testsuite/gnatprove/tests/Q929-022__flow_illegal_varinput/test.out
Update tests.

Change-Id: Ie46a8d6d2383765c45511f01c2910fe9d6434a43