R216-019 Crash on type used in refined postcondition
When a Refined_Post was used on a subprogram whose body is not in
SPARK, while its spec is in SPARK, this could lead to a crash.
Now fixed.

* gnat2why/why/gnat2why-expr.adb
* gnat2why/why/gnat2why-subprograms.adb
* gnat2why/why/why-atree-modules.adb
Protect all uses of the refined postcondition by checking not only
that it exists, but also that the body of the subprogram is in SPARK.

* testsuite/gnatprove/tests/R216-019__instance_fun/test.opt
Test passes now.

Change-Id: Ia5e1f78750171410ce375b7a11879961a3520c26