R216-019 assert for preventing wrong access to Refined_Post
The crash fixed on this TN happened because GNATprove was accessing
Refined_Post of subprograms whose body was not annotated with SPARK_Mode =>
On. We could detect such access by an assertion within the Find_Contracts
routine. This patch add such an assertion, so that a similar problem is less
likely to happen again.

* gnat2why/spark/spark_util-subprograms.adb
(Find_Contracts): add assertion (it is easier to check in body than in Pre).

* gnat2why/spark/spark_util-subprograms.adb
(Checking_Of_Refined_Post): this routine is only executed for subprograms
 with bodies in SPARK; this is now asserted in the Pre contract and not
 checked in production builds.

(Translate_Expression_Function_Body): refactor to avoid repeated calls to
 Entity_Body_In_SPARK and Has_Contracts.

Other changes merely flip the order of Entity_Body_In_SPARK and
Has_Contracts, so that the latter is only called when the Refined_Post is
visible.

Change-Id: I47f12d9d956b4fca4d491ec51ca27c5546f9d99a