R213-039 Minor change for stability of sumrange
Removed altergo from the session. Adding --replay and moved
--no-axiom-guard to the project file.
Changing replay functions where needed. Added manual boolean value to
test.py using replay and made a case on this boolean in update-session.
If the boolean is true, sessions should not be removed.

* testsuite/gnatprove/tests/QA11-003__itp_sumrange/*
Update test for stability.

* testsuite/gnatprove/tests/*/test.py
Adding manual boolean in test using replay.

* testsuite/gnatprove/update-session-test
Use boolean contains_manual_proof to not remove session when there is a
manual proof.
Adapted stuff so that pre-commit-check and pep8 are happy.


Change-Id: Ie983f0cfe7922b38f1003bf7e08e6a163f22f3f2