R205-008 better slocs for constant literals in Global contracts
On this TN we already fixed searching for N_Numeric_Or_String_Literal in the
Depends/Refined_Depends/Initializes contract. This patch does the same for
Find_Global routine, which searches in the Global/Refined_Global
contracts (or in the Initializes, when called for a package entity).

* gnat2why/flow/flow-analysis.adb
(Find_Global): find globals represented by a rewritten constant.

* testsuite/gnatprove/tests/O414-015__flow_constants_in_contracts...
* testsuite/gnatprove/tests/TU__types_and_subtypes_illegal/test
Update slocs in test outputs.

Change-Id: Ia999c4f2868770abe11531bab612a5cfd0d86bff