R209-002 avoid crossing SPARK bondaries in Root_Record_Component
When searching for the root component associated to a discriminant,
we use the Corresponding_Discriminant link to track renamed
discriminants. This can lead us to cross the SPARK boundary if the
full view of the record type is not in SPARK.

* gnat2why/spark/spark_util.adb
(Root_Record_Component): Only use the Corresponding_Discriminant
 link if it links to a type which is in SPARK.

* tests/R209-002__container_iteration/test.opt
Remove; test is now executed as expected (without output).

Change-Id: I972e3b9032dcce4e5ed62880698b376b5983a6d6