QC08-004 mark Standard package together with other units
There was no reason for marking the Standard package and marking of other
units to be interleaved with rewriting the AST and registering the
Entity_Id<->Entity_Name mappings.

Now the Standard package is marked just like other packages.

Note: marking of the standard entities is likely to be changed anyway as
part of the top-down marking, but at least now the code is cleaner and
easier to adapt to the new Is_Visible machinery.

* gnat2why/why/gnat2why-driver.adb
* gnat2why/spark/spark_definition.adb
Move call to Mark_Standard_Package from GNAT_To_Why to
Mark_Compilation_Unit; tool behaviour remains as it was.

Change-Id: I9936d339412ca08478a48d7875e7436be04e81e7