R206-025 Initial condition of withed units on subprograms
Do not assume the initial condition of withed units for compolation
units which are subprograms. Theoretically, they can be assumed to
hold for the Main subprogram but we currently have no way to decide
which subprogram is the Main.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_VCs_For_Subprogram): Never assume initial condition of
 withed units.

* tests/R206-025__initial_condition/test.opt
Remove, test is enabled.

* tests/R206-025__initial_condition/test.out
No soundness bug anymore.

* tests/MB21-036__mainpre/test.out
Provability regression, we no longer assume the initial condition
of withed units for the main subprogram.

Change-Id: I8fe91dc455f753a93bca3a168366230ddc835dee