P805-023, QC08-004 don't load System from marking
Loading the System unit while marking was suspicious from the start (because
it is unclear whether Sem.Walk_Library_Items is prepared for that). Now it
became a problem, because it interferes with new Is_Visible machinery.

New Is_Visible relies on traversing all units BEFORE marking (because
marking relies on visibility). If we pull the System unit only at marking,
we will crash in proof when asking for visibility of the System entities,
because we know nothing about visibility of the System entities. This patch
simply pulls the System unit earlier (but still only when necessary).

* gnat2why/spark/spark_definition.adb
(Mark_Entity): add assertion to make sure that the System unit is already
 loaded.

(Mark_Subprogram_Body): only mark System.Default_Priority expression when
 analyzing a main-like unit, not when WITHin such a unit; add assertion just
 like in the Mark_Entity; clarify comment.

* gnat2why/why/gnat2why-driver.adb
(Gnat_To_Why): load the System unit before marking.

Change-Id: Id8ed7e238fea11c343e64f7e57163896b8b8c9da