R221-027 Correct bounds for multidimensional array aggregates
For record aggregates, gnat2why generates an axiom describing the
values contained in the aggregate (see Transform_Aggregate). This
axiom also gives the values of the bounds of record aggregates of a
dynamic type. For multidimensional arrays, the assumptions for the
values of the different ranges all mention bounds of the first
range. This is unsound as it gives different values for the same
constant whenever the bounds of the ranges are different.

* gnat2why/why/gnat2why-expr.adb
(Transform_Aggregate): Use the appropriate bounds in the aggregate
 axiom.

* tests/R221-027__multidim_aggr
Enable the test.

Change-Id: Ib0dfb5e5ae9ff92057dde34e0fba3a9f837c8858