R220-054 Support more values of small for fixed-point types
So far, only the values of fixed-point smalls that were negative powers
of 2 and 10 were supported. This is now extended to all smalls which are
reciprocals of integers. This is possible because the guarantees of Ada RM
about precise results extends to those values of small, and because the
current axiomatization is valid for such values.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Allow more values of small.

* share/spark/theories/ada__model.mlw
(Fixed_Point_Base): Update comments.

* testsuite/gnatprove/tests/KB28-016__over/test.out
* testsuite/gnatprove/tests/M823-017__fixed_points/test.out
Update output.

* testsuite/gnatprove/tests/M823-017__fixed_points/ordinary_fixed.adb
* testsuite/gnatprove/tests/M823-017__fixed_points/ordinary_fixed.ads
* testsuite/gnatprove/tests/M823-017__fixed_points/test.out
* testsuite/gnatprove/tests/M823-017__fixed_points/test.py
Update test to have a version with newly supported small.

* tests/R220-054__fixed_point_small/fp_test.adb
* tests/R220-054__fixed_point_small/fp_test.ads
* tests/R220-054__fixed_point_small/main.adb
* tests/R220-054__fixed_point_small/test.out
* tests/R220-054__fixed_point_small/test.py
New test.

Change-Id: Ibd38588983fb075ee99ab01770e4b2fca7c745a5