Q413-009 Remove the requirement of only one prover for manual provers
This removes the requirement of using only one prover when doing manual
proof (with coq). This is needed because we need to build the directory
why3_libs containing the Coq file when in --replay mode. In --replay mode
we want several provers *and* Coq.
This also switch Q405-006__coq_lemma_library to replay mode.

* gnatprove/configuration.adb
Several prover can be called with coq.

* include/proof/sessions/spark_float*/why3session.xml
Updated session for lemma library

* testsuite/gnatprove/tests/P405-006__coq_lemma_library/test.*
Adapted test

Change-Id: I0d202cdf7ff6f6a8390933e607edd5967021fabf