R227-009 support intrinsic operators on private types
Intrinsic operators are not rewritten by the frontend as function
calls even on private types whose full view is a scalar type. If
the full view of the private types is not in SPARK, we need to
translate them as function calls in Why.

* gnat2why/why/gnat2why-expr.adb
(Compute_Call_Args): Support being called on operators.
(Transform_Comparison): Put back assertion ensuring that comparison
 operators on record types are equalities or disequalities.
(Transform_Expr): Intrinsic operators on private types are
translated as function calls if their entity is a function.
(Transform_Function_Call): Support being called on operators.

* testsuite/gnatprove/tests/R227-009__intrinsic_op
Enable test.

Change-Id: I3a2a111744a0f28e271541bd44327077da28f224