R219-019 Add a args_time argument in gnatprove.conf
We changed the driver so that the command when steps are used is unchanged
but, for counterexample run, when steps are not used, a time_limit
argument is given.
In practice, we transform "args" from gnatprove.conf into a generic
argument for the given prover which is either concatenated with args_time
or args_steps. In practice, steps and timelimit won't be usable at the
same time.

* gnatprove/gnatprove.adb
(Build_Prover): Change the parsing to add the args_time argument (which is
empty for normal run).

* share/spark/config/gnatprove.conf:
Add an args_time argument wherever necessary.

* testsuite/gnatprove/tests/N731-010__aggregate/test.out
Update diff in counterexample

Change-Id: Ib46ff0a5eecffe26ff20fff6c7af2d7a7e21422e