R222-012 use to_rep instead of of_rep in axioms for aggregates
Array aggregates are translated as abstract logic functions taking
values of the aggregate as parameters. An axiom is generated which
links the result of this function with the values of the aggregate.
This axiom used to express these equalities using equality on the
abstract view of the element type. Now the equality is done on the
split view. For this to be sound, guards have been added to force
the value parameters to be in bound. This does not seem to help
proof but it produces slightly better counterexamples.

* gnat2why/why/gnat2why-expr.adb
(Transform_Aggregate): Use equality on base type for array elements
 in the defining axiom of the aggregate. To preserve soundness, add
 a guard ensuring that values for the aggregate are in bound. Move
 definition axiom to axiom module as it now needs to reference
 dynamic properties which are defined in completion modules of
 types.

* gnat2why/why/why-atree-modules.adb
(E_Axiom_Module): Support the construction of axiom modules on
 aggregate nodes.
(Insert_Extra_Module): New parameter Is_Axiom to allow adding an
 extra axiom module.

* gnat2why/why/why-inter.adb
(Dispatch_Entity_Completion): Support dispatching the completion of
 aggregate nodes.

* testsuite/gnatprove/tests/*/test.out
Update test outputs. Mostly changes in counterexamples and proof
regressions from CVC4 that are regained by either altergo or Z3.

Change-Id: I185d336ce2e036d284eb8444599150e9779e4bdb