do not mark object's initial value if its types are not in SPARK
This patch reduces the number of error messages from marking and seems in
the spirit of the top-down marking.

An immediate motivation was to understand checks for the SPARK status of
full views when a box notation is used in aggregate expressions (thought
this notation can be used not only in object initialization expressions).

(no-tn-check)

* gnat2why/spark/spark_definition.adb
(Mark_Object_Entity): return on violations in the object's types.

Change-Id: I0328d53344f6cd9af4ca69d8d2de6ea8e025f8e4