R307-019 support 'Old attribute on protected components
Old attribute can be called on protected components and Part_Of
variables, for example in refined postconditions of protected
subprograms. We no longer crash on this construct.

* gnat2why/why/gnat2why-expr.adb
(Name_For_Old): Entities of protected components are not stored in
 the entity map. Call Type_Of_Node instead of Why_Type_Of_Entity to
 support this case.

* gnat2why/why/gnat2why-subprograms.adb
(Compute_Dynamic_Property_For_Effects): For subprograms which are
 declared in a protected object, there is a special self parameter
 used to represent the protected object. Handle this parameter
 specifically.

* testsuite/gnatprove/tests/R307-019__protected_old
Extend test to check part of variables, effective translation of
the old attribute, and internal calls to protected procedures with
refined postconditions.

Change-Id: I5494dd02412c3d1970bb5a1327265d96132fc905