R309-018 intrinsic operators of private types in predicates
We currently crash on intrinsic operators on private types whose
full view is not in SPARK when they are in the EW_Pred domain.
Indeed, we try to translate the underlying function call in the
predicate domain, causing errors when translating parameters which
are not standard booleans.

* gnat2why/why/gnat2why-util.adb
(Is_Private_Intrinsic_Op): returns True on intrinsic operator on
 private types.

* gnat2why/why/gnat2why-expr.adb
(Transform_Expr): Do not translate boolean operators as predicates
 if they are private.

* tests/R309-018__real_time_ops/
File test case.

Change-Id: I426bbe710231ddf9c9722317e5e85dc45acc1e5e