correct failing assertion on classwide types with predicates
Type_Needs_Dynamic_Invariant is a function used only for purpose of
emitting better info messages (we only emit a message about
implicit function contract not availlable if there really is an
implicit function contract). To keep it syncronized with the actual
implementation of Dynamic_Invariant, it is also used as a post of
Compute_Dynamic_Invariants. This postcondtion was failing on
classwide views of types which need dynamic invariants because of a
bad handling of this case in Type_Needs_Dynamic_Invariant.

(no-tn-check)

* gnat2why/why/gnat2why-util.adb
(Type_Needs_Dynamic_Invariant): Go to the specific view of
 classwide types.

Change-Id: Iaa16e58242a3fa9ed0c44c029c41a80441eb7aac