fix initialization of flow record info
When flooding the current component, whose entity is kept in the Ptr
variable, we should not modify the Ptr itself.

In particular, when we modify Ptr when iterating over Out_Neighbours, then
we will iterate over In_Neighbours of some other component. This looks like
a bug, but appears to have no impact on the results.

Also, the current component (Ptr) should be removed from Work_List after
iterating over its in/out-neighbours, to prevent its repeated processing
(as it could be connected to itself, i.e. be its own neighbour).

(no-tn-check)

* gnat2why/flow/flow_utility.adb
(Initialize): don't overwrite the Ptr variable.

Change-Id: I37de77f813cc8b2567b76c33231c65505fd6b673