PA07-031 better abstraction of GNAT tree
The aim of this ticket is to shadow the atree of GNAT frontend and
to replace it with a SPARK specific version. The aim is to abstract
away some of the complexity of the translation to why which comes
from using routines which are not SPARK specific.

* gnat2why/why/spark_atree.ads
* gnat2why/why/spark_atree-entities.ads
New file shadowing frontend routines.

* gnat2why/why/gnat2why-expr-loops-inv.adb
* gnat2why/why/gnat2why-expr-loops.adb
* gnat2why/why/gnat2why-expr.adb
* gnat2why/why/why-gen-arrays.adb
* gnat2why/why/why-gen-records.adb
Use routines from spark_atree.

Change-Id: I3178c6827ddb487a27d934693f753011ad680845