Updated git submodules
Project: why3 master a2638f969331ac82c7ad20a691b022dfdaee9438

R312-016 remove call of adapt_limit function

This function is not needed in gnatwhy3. It is used to compute the limits
for a replay call, but gnatwhy3 does it itself already. Also, this
function should not be called when the prover call is not a replay.

* controller_itp.ml
(schedule_proof_attempt): remove computation of adapted limit

Change-Id: Ie4d020eb32a170e0b00ac5a73deab8dfb72ddd53