Updated git submodules
Project: why3 master eea3d12e144f830986fa23824bb27a91d99d33ac

Q915-037 Split during counterexamples pass

This commit tries to implement an algorithm of splitting for getting
counterexamples. This was designed to not impact the existing proof pass.
When one of the goal of an objective (=Spark check) cannot be proved with
provers that were queried by the user, the result status Counterexample
is given. All the other goals related to this objective are abandoned:
upon return, they won't trigger getting a counterexample because
counterexample has already be triggered by one of the goal.
During counterexample pass, the first goal is queried for counterexample.
If this proves the goal, we stop and don't return counterexamples.
If not, If it is fully split, we return this counterexample and stop any
other counterexamples getting.
        If not, we split using transformation transform_ce, and we add the
newly split goals into the set of goals to try for counterexample. If
there was a model, we save it as a temporary model (to be used if the
splitted goal do not generate one).
Note that counterexamples are run one by one contrary to regular proof
which tries to run Gnat_config.parallel provers per objective (in theory,
we can imagine making this number arbitrarily large with a large split and
carefully managed time to prove on each goal).
Note that this also *save the trace* in all modes now.

* src/gnat/gnat_main.ml
(handle_vc_result): Adding has_model to naively detect the absence of
counterexample. Adding ce to change register_result behavior during
counterexample pass. When Counter_Example is given, tries to run prover
for next counterexample.
(interpret_result): Adding has_model which corresponds to the
presence/absence of model.
(create_manual_or_schedule): Specific behavior for counterexample pass.
(actually_schedule_goal): inlined
(schedule_goal): space added/removed.
(report_messages): Instead of getting a random proofattempt, we now get
the one that generated counterexamples. Behavior without counterexamples
is not affected.
The trace is now saved in *all* mode to allow filtering of
counterexamples.

* src/gnat/gnat_objectives.ml
(objective_rec): Adding to_be_counterexampled to record the remaining
goals on which to try to get counterexamples. Add temporary_ce to record
a counterexamples before splitting a goal. Adding definitive_ce to record
a counterexample that cannot be split more (this stop all counterexample
getting).
(add_to_objective_ce): Same as add_to_objective but for
to_be_counterexampled.
(find_ce): Find the most relevant counterexample in an objective.
(next_ce): Same as next but only return one of to_be_counterexampled set.
(transform_ce): transform_ce is the *only* split transformation that will
be used for ce.
(is_full_split_goal_ce): Check if transform_ce was already applied.
(subsubgoals_ce): returns the subgoals of transform_ce was applied, []
otherwise.
(further_split_ce): Applies transform_ce if possible.
(register_result): Two cases depending if ce is true or not. When ce is
false the former case should be unchanged. When ce is true, the algorithm
described is implemented.
(session_find_unproved_goal): It is possible that goals that do not belong
to an objective are found. We add a check that it belongs to goalmap. It
can probably also happens with manual proofs.


Change-Id: Ic95c9058ee2eb292bc8cf270bfbb8506d8f115b9