R122-044 Distinguish exit status between errors and check messages
A recent change caused GNATprove to return with a non-zero exit status
when unproved check messages were issued. This was convenient for some
users, but not all of them. With this new patch, we revert to the previous
default where unproved check messages do not cause a non-zero exit status.
Instead, GNATprove has a new switch --checks-as-errors that causes
unproved check messages to be treated as errors.

* docs/ug/en/source/how_to_run_gnatprove.rst
Document new behavior.

* gnatprove/configuration.adb
* gnatprove/configuration.ads
Add support for switch --checks-as-errors.

* gnatprove/gnatprove.adb
(Generate_SPARK_Report): Consider checks as errors only when switch
 --checks-as-errors is used.

* share/spark/help.txt
Add switch --checks-as-errors.

* testsuite/gnatprove/tests/QB03-022__exit_status/test.out
* testsuite/gnatprove/tests/QB03-022__exit_status/test.py
Update test to use new switch.

Change-Id: Ic6168e00fb7f365129411405c050f97f5ee32c5f