R314-016 remove special handling of discriminants for defaults
We used to specifically store values for discriminants entities of
record base types when translating default values of components,
because default values are defined in base types only and therefore
would refer to discriminants enties of base types instead of their
own. However, this is not longer needed since the handling of
discriminants in the Symbol_Table has been changed so that they are
normalized.

* gnat2why/why/gnat2why-expr.adb
(Compute_Default_Check): Remove special case for discriminants of
 base types.

Change-Id: I751048b6f0f825ad967186d2f760930b0fc9e885