R216-008 Remove spurious checks on package bodies not in SPARK
Library-level package bodies not marked in or out of SPARK should be
traversed for subprograms/packages that would have SPARK_Mode=>On, but
they should not themselves be considered in SPARK. In particular,
their elaboration code should not be analyzed.

* gnat2why/spark/spark_definition.adb
(Mark_Package_Body): Only consider body in SPARK if SPARK_Mode=>On
 enforced on package body.

* testsuite/gnatprove/tests/MA22-019__spark_mode/p.adb
* testsuite/gnatprove/tests/P630-025__flow_spec_auto/p.ads
Update test.

* testsuite/gnatprove/tests/R216-008__pkg_elab_mode/test.out
Test passes now.

Change-Id: Ib035661e8ae16476a01588d126674a213af3a722