P803-008 Fix crash in Check_Elaborate_Body
The code in Check_Elaborate_Body used to crash when seeing an Entity_Name
variable. This is now fixed. Public testsuite not affected but private
testsuite will be.

* gnat2why/flow/flow-analysis.adb
(Check_Elaborate_Body): we are now iterating over Parse_Initializes to
check for variables that we claim to initialize. We need to check that
the variable in question is declared in the spec of the package.

Change-Id: I545a2c4504edf0956ebcf1cc18b5158e4fdb1680